Moved Mock files, added stop test to TabTests
