+++ firefox-ios/firefox-ios-tests/Tests/ClientTests/Mocks/MockURLProtocol.swift
@@ -1,0 +1,40 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+
+class MockURLProtocol: URLProtocol {
+    static var response: ((HTTPURLResponse, URLRequest) -> Void)?
+    static var data: Data?
+
+    override class func canInit(with request: URLRequest) -> Bool {
+        return true
+    }
+
+    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
+        return request
+    }
+
+    override func startLoading() {
+        let response = HTTPURLResponse(
+            url: request.url!,
+            statusCode: 200,
+            httpVersion: nil,
+            headerFields: nil
+        )!
+        Self.response?(
+            response,
+            request
+        )
+        if let data = Self.data {
+            client?.urlProtocol(self, didLoad: data)
+        }
+
+        client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
+        client?.urlProtocolDidFinishLoading(self)
+    }
+
+    override func stopLoading() {
+    }
+}
